+++ src/collective/solr/tests/test_doctests.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 from collective.solr.testing import COLLECTIVE_SOLR_FUNCTIONAL_TESTING
-from collective.solr.tests.utils import pingSolr
 from plone.testing import layered
 from unittest import TestSuite
 import doctest
@@ -11,20 +10,19 @@ optionflags = (doctest.ELLIPSIS | doctest.NORMALIZE_WHITESPACE
 
 def test_suite():
     suite = TestSuite()
-    if pingSolr():
-        for testfile in [
-            'errors.txt',
-            'configlet.txt',
-            'search.txt',
-            'conflicts.txt',
-            'facets.txt',
-            'dependencies.txt',
-            'collections.txt',
-        ]:
-            doc_suite = doctest.DocFileSuite(testfile,
-                                             package='collective.solr.tests',
-                                             optionflags=optionflags)
-            layer = layered(doc_suite,
-                            layer=COLLECTIVE_SOLR_FUNCTIONAL_TESTING)
-            suite.addTest(layer)
+    for testfile in [
+        'errors.txt',
+        'configlet.txt',
+        'search.txt',
+        'conflicts.txt',
+        'facets.txt',
+        'dependencies.txt',
+        'collections.txt',
+    ]:
+        doc_suite = doctest.DocFileSuite(testfile,
+                                         package='collective.solr.tests',
+                                         optionflags=optionflags)
+        layer = layered(doc_suite,
+                        layer=COLLECTIVE_SOLR_FUNCTIONAL_TESTING)
+        suite.addTest(layer)
     return suite